seperate employee and customer views
