+++ src/components/nav/CustomerNavBar.js
@@ -1,0 +1,26 @@
+import { Link, useNavigate } from "react-router-dom"
+import "./NavBar.css"
+
+export const CustomerNavBar = () => {
+    const navigate = useNavigate()
+
+    return (
+        <ul className="navbar">
+            <li className="navbar__item active">
+                <Link className="navbar__link" to="/locations">Locations</Link>
+            </li>
+            <li className="navbar__item active">
+                <Link className="navbar__link" to="/products">Products</Link>
+            </li>
+            <li className="navbar__item active">
+                <Link className="navbar__link" to="/find-candy">Find Candy</Link>
+            </li>
+            <li className="navbar__item navbar__logout">
+                <Link className="navbar__link" to="" onClick={() => {
+                    localStorage.removeItem("kandy_user")
+                    navigate("/", {replace: true})
+                }}>Logout</Link>
+            </li>
+        </ul>
+    )
+}
