+++ src/components/nav/NavBar.js
@@ -1,23 +1,19 @@
-import { Link, useNavigate } from "react-router-dom"
-import "./NavBar.css"
+import { CustomerNavBar } from "./CustomerNavBar"
+import { EmployeeNavBar } from "./EmployeeNavBar"
 
 export const NavBar = () => {
-    const navigate = useNavigate()
 
-    return (
-        <ul className="navbar">
-            <li className="navbar__item active">
-                <Link className="navbar__link" to="/locations">Locations</Link>
-            </li>
-            <li className="navbar__item active">
-                <Link className="navbar__link" to="/products">Products</Link>
-            </li>
-            <li className="navbar__item navbar__logout">
-                <Link className="navbar__link" to="" onClick={() => {
-                    localStorage.removeItem("kandy_user")
-                    navigate("/", {replace: true})
-                }}>Logout</Link>
-            </li>
-        </ul>
-    )
+    const localKandyUser = localStorage.getItem("kandy_user")
+    const kandyUserObject = JSON.parse(localKandyUser)
+
+    if (kandyUserObject.staff) {
+        //return staff nav
+        return <EmployeeNavBar />
+    }
+
+    else {
+        //return customer nav
+        return <CustomerNavBar />
+    }
+
 }
