+++ src/components/views/ApplicationViews.js
@@ -1,25 +1,19 @@
-import { Outlet, Route, Routes } from "react-router-dom"
-import { LocationList } from "../locations/LocationList"
-import { ProductForm } from "../products/ProductForm"
-import { ProductList } from "../products/ProductList"
-
+import { CustomerViews } from "./CustomerViews"
+import { EmployeeViews } from "./EmployeeViews"
 
 export const ApplicationViews = () => {
-	return (
-		<Routes>
-			<Route path="/" element={
-				<>
-					<h1>Candy Corner Sweets Shop</h1>
-					<div className="subheading">Your wholesale sweet tooth soothers for almost 35 years</div>
-			
-					<Outlet />
-				</>
-			}>
-				<Route path="/locations" element={<LocationList />} />
-				<Route path="/products" element={<ProductList />} />
-				<Route path="/product/create" element={<ProductForm />} />
-			</Route>
 
-		</Routes>
-	)
+    const localKandyUser = localStorage.getItem("kandy_user")
+    const kandyUserObject = JSON.parse(localKandyUser)
+
+    if (kandyUserObject.staff) {
+        //return staff views
+        return <EmployeeViews />
+    }
+
+    else {
+        //return customer views
+        return <CustomerViews />
+    }
+
 }
